Only modify the buffer in the list and repeat mappings if it would cause an error
The user may decide to cancel the selection of previous insertions (via <Enter> or <Esc>), and then a previous repeat of a previous insertion would be clobbered, as the setline() has incremented b:changedtick without a new repeat from the mapping.
Likewise, a repeat may fail because the user supplied a non-existing named recall, resulting in an error message and no new repeat. The user then expects to keep the previous (successful) repeat (and may try repeating again with a different register).