Add separate PrevInsertComplete#Recall#RecallRepeat()
And reset the count if the actual register differs from the original register. Due to the dual nature of the count (last insertion number vs. multiplier), we cannot simply use the count we've gotten; only if the original register remains on repeat. If the user overrides the register ("r.), we need to reset the count, to avoid that a previous recall of the 4'th last previous insertion (4q<A-a>, with the default register) now inserts named recall "d 4 times (on "d.)!